// generated/src/app/api-models/member-address.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class MemberAddress {
  seqMembId: number;
  addressType: string;
  insertDatetime: Date;
  insertUser: string;
  updateDatetime: Date;
  updateUser: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  state: string;
  zipCode: string;
  county: string;
  country: string;
  homePhoneNumber: string;
  fax: string;
  email: string;
  busPhoneNumber: string;
  beneficiaryRelCode: string;
  beneficiaryDob: Date;
  beneficiaryGender: string;
  memAddrUserDefined1: string;
  memAddrUserDefined2: string;
  memAddrUserDate1: Date;
  memAddrUserDate2: Date;
  lastName: string;
  firstName: string;
  mobilePhone: string;
  insertProcess: string;
  updateProcess: string;
  securityCode: string;
}